export Parse function to use as a library
