+++ .gitignore
@@ -23,4 +23,4 @@ _testmain.go
 *.test
 *.prof
 
-epp
+/bin/epp
